third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-block-indent-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
This test verifies applied indentation whenever you hit enter in "{|}" or type in "}" while inside opened block.

{} {}
  {}
  {
    
    {
        


Running: testSimpleCollapsedBlockExpanding
{
..|
}.{}
..{}
..{
....
....{
........


Running: testMulticursorCollapsedBlockExpanding
{
..|
}.{
..|
}
..{
....|
..}
..{
....
....{
........


Running: testMulticursorCollapsedBlockNotExpanding
{
|}.{}
..
..|{}
..{
....
....{
........


Running: testSingleCursorClosingBracketIndent
{}.{}
..{}
..{
..}|....
....{
........


Running: testMulticursorClosingBracketIndent
{}.{}
..{}
..{
..}|
....{
....}|


Running: testMulticursorClosingBracketIndentNotExecuted
{}.{}
..{}
..{
....}|
....{}|
........
